move isOnline state to CalendarApp from useRemoteSync
